Moved exitstatus back out
