+++ src/test/kotlin/com/manimdsl/semanticanalysis/InvalidSemanticTests.kt
@@ -1,7 +1,7 @@
 package com.manimdsl.semanticanalysis
 
-import com.manimdsl.frontend.ExitStatus
-import com.manimdsl.frontend.ManimDSLParser
+import com.manimdsl.ExitStatus
+import com.manimdsl.ManimDSLParser
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
 import org.junit.jupiter.api.*
